Room Effects WIP
I've implemented the tier 1 effects of the shrine, study, storage, and refinery. I need to add an "Assign Unit" button to the refinery, making it only usable while there is a worker in the room. The study is the only room that doesn't require a worker. Need to add unit removal from rooms, currently units just stop at the robot if the storage is full.